refactor(utils): add helper
