+++ packages/plugin-lighthouse/src/lib/utils.ts
@@ -1,5 +1,5 @@
 import type { CliFlags } from 'lighthouse';
-import { objectToCliArgs, toArray } from '@code-pushup/utils';
+import { objectToCliArgs } from '@code-pushup/utils';
 import { LIGHTHOUSE_REPORT_NAME } from './constants';
 
 type RefinedLighthouseOption = {
@@ -47,58 +47,4 @@ export function getLighthouseCliArguments(
   }
 
   return objectToCliArgs(argsObj);
-}
-
-export class AuditsNotImplementedError extends Error {
-  constructor(list: WithSlug[], auditSlugs: string[]) {
-    super(
-      `audits: "${auditSlugs
-        .filter(slug => !list.some(a => a.slug === slug))
-        .join(', ')}" not implemented`,
-    );
-  }
-}
-
-export function filterByAuditSlug<
-  T extends {
-    refs: WithSlug[];
-  },
-  S extends WithSlug['slug'] = T['refs'][number]['slug'],
->(groups: T[], auditSlugs: S | S[]): T[] {
-  const slugs = toArray(auditSlugs);
-  if (slugs.length === 0) {
-    return groups;
-  }
-  return (
-    groups
-      // filter out groups that have no audits includes from onlyAudits (avoid empty groups)
-      .filter(group => group.refs.some(({ slug }) => slugs.includes(slug as S)))
-      .map(group => {
-        const groupsRefs = group.refs.filter(({ slug }) =>
-          slugs.includes(slug as S),
-        );
-
-        return {
-          ...group,
-          refs: groupsRefs,
-        };
-      })
-  );
-}
-
-export type WithSlug = { slug: string };
-
-export function filterBySlug<
-  T extends WithSlug,
-  S extends WithSlug['slug'] = WithSlug['slug'],
->(list: T[], auditSlugs: S | S[]): T[] {
-  const slugs = toArray(auditSlugs);
-  if (slugs.length === 0) {
-    return list;
-  }
-  if (slugs.some(slug => !list.some(wS => wS.slug === slug))) {
-    throw new AuditsNotImplementedError(list, slugs);
-  }
-
-  return list.filter(({ slug }) => slugs.includes(slug as S));
 }
