+++ packages/plugin-lighthouse/src/lib/utils.unit.test.ts
@@ -1,9 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import {
-  filterByAuditSlug,
-  filterBySlug,
-  getLighthouseCliArguments,
-} from './utils';
+import { getLighthouseCliArguments } from './utils';
 
 describe('getLighthouseCliArguments', () => {
   it('should parse valid options', () => {
@@ -24,63 +20,5 @@ describe('getLighthouseCliArguments', () => {
         '--chromeFlags="--headless=new --user-data-dir=test"',
       ]),
     );
-  });
-});
-
-describe('filterBySlug', () => {
-  it('should return the given list if no slugs are given', () => {
-    const list = [{ slug: 'a' }, { slug: 'b' }, { slug: 'c' }];
-    expect(filterBySlug(list, [])).toEqual(list);
-  });
-
-  it('should filter if slugs is a string', () => {
-    const list = [{ slug: 'a' }, { slug: 'b' }, { slug: 'c' }];
-    expect(filterBySlug(list, 'a')).toEqual([{ slug: 'a' }]);
-  });
-
-  it('should filter if slugs is an array', () => {
-    const list = [{ slug: 'a' }, { slug: 'b' }, { slug: 'c' }];
-    expect(filterBySlug(list, ['a'])).toEqual([{ slug: 'a' }]);
-  });
-});
-
-describe('filterByAuditSlug', () => {
-  it('should return the given list if no slugs are given', () => {
-    const list = [
-      {
-        refs: [{ slug: 'a' }, { slug: 'b' }, { slug: 'c' }],
-      },
-    ];
-    expect(filterByAuditSlug(list, [])).toEqual([
-      {
-        refs: [{ slug: 'a' }, { slug: 'b' }, { slug: 'c' }],
-      },
-    ]);
-  });
-
-  it('should filter if slugs is a string', () => {
-    const list = [
-      {
-        refs: [{ slug: 'a' }, { slug: 'b' }, { slug: 'c' }],
-      },
-    ];
-    expect(filterByAuditSlug(list, 'a')).toEqual([
-      {
-        refs: [{ slug: 'a' }],
-      },
-    ]);
-  });
-
-  it('should filter if slugs is an array', () => {
-    const list = [
-      {
-        refs: [{ slug: 'a' }, { slug: 'b' }, { slug: 'c' }],
-      },
-    ];
-    expect(filterByAuditSlug(list, ['a'])).toEqual([
-      {
-        refs: [{ slug: 'a' }],
-      },
-    ]);
   });
 });
