+++ packages/utils/src/lib/transform.ts
@@ -1,4 +1,5 @@
 import { platform } from 'node:os';
+import { Audit, GroupRef } from '@code-pushup/models';
 
 export function toArray<T>(val: T | T[]): T[] {
   return Array.isArray(val) ? val : [val];
@@ -156,3 +157,56 @@ export function toOrdinal(value: number): string {
   return `${value}th`;
 }
 /* eslint-enable no-magic-numbers */
+
+export class AuditsNotImplementedError extends Error {
+  constructor(list: { slug: string }[], auditSlugs: string[]) {
+    super(
+      `audits: "${auditSlugs
+        .filter(slug => !list.some(a => a.slug === slug))
+        .join(', ')}" not implemented`,
+    );
+  }
+}
+
+export function filterByAuditSlug<
+  T extends {
+    refs: GroupRef[];
+  },
+  L extends T[],
+  S extends L[number]['refs'][number]['slug'],
+>(groups: L, auditSlugs: S | S[]): L {
+  const slugs = toArray(auditSlugs);
+  if (slugs.length === 0) {
+    return groups;
+  }
+  return (
+    groups
+      // filter out groups that have no audits includes from onlyAudits (avoid empty groups)
+      .filter(group => group.refs.some(({ slug }) => slugs.includes(slug as S)))
+      .map(group => {
+        const groupsRefs = group.refs.filter(({ slug }) =>
+          slugs.includes(slug as S),
+        );
+
+        return {
+          ...group,
+          refs: groupsRefs,
+        };
+      }) as L
+  );
+}
+
+export function filterBySlug<T extends Audit, L extends T[]>(
+  list: L,
+  auditSlugs: L[number]['slug'][] | L[number]['slug'],
+): L {
+  const slugs = toArray(auditSlugs);
+  if (slugs.length === 0) {
+    return list;
+  }
+  if (slugs.some(slug => !list.some(wS => wS.slug === slug))) {
+    throw new AuditsNotImplementedError(list, slugs);
+  }
+
+  return list.filter(({ slug }) => slugs.includes(slug)) as L;
+}
